#161 Exported aggregations to a separate file
